Make libs updates public
Forgot to move all the changes into the library folders.
